feat: enable syntax highlighting for Asciidoc
Currently embedding inline styles.  Not ideal but couldn't figure out
how to link.  The asciidoctor linkcss attribute [1] can be used to do
this but the <link> may be in the <head> generated/stripped by
asciidoctor.

I also tried the Rouge highlighter but couldn't get it to work.
Probably because of a stylesheet issue.  I don't think there is a way to
embedded Rouge styles like you can with Pygments.

I tried using a different theme other than default/friendly but anything
with a non-gray background had rendering issues where the test
background would be gray but the rest of the div background would be the
right theme color.

[1]: https://docs.asciidoctor.org/asciidoctor/latest/html-backend/source-highlighting-stylesheets/#pygments